move evm specific pages into /evm
