+++ src/components/Checkbox.jsx
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React from "react";
 import styled from "styled-components";
 
 const CheckboxWrapper = styled.div`
